Concordance Index

Compute concordance index (C-index or C-statistic) that allows weights for right-censored survival data. For example, Asano and Hirakawa (2017) proposed cure status weighting for cure models, which reduces to Harrell's C-index if weighs are all ones.

cIndex(time, event = NULL, risk_score, weight = NULL)

Arguments

time
A numeric vector for observed times

event
A numeric vector for event indicators. If it is NULL (by default) or NA, event will be treated all as ones and the function will compute concordance index for uncensored survival data.

risk_score
A numeric vector representing the risk scores of events.

weight
A optional numeric vector for weights. If it is NULL (by default) or NA, equal weights will be used.

Value

A named numeric vector that consists of

index: the concordance index.

concordant: the number of concordant pairs.

comparable: the number of comparable pairs.

tied_tisk: the number of comparable pairs having tied risks.

Details

Let \(r_i\), \(t_i\), and \(\delta_i\) denote the risk score, observed time, and event indicator of \(i\)-th subject. The pair of \((t_i,\delta_i)\) and \((t_j,\delta_j)\), where \(i<j\), are defined to be comparable if \(t_i<t_j,\delta_i=1\) or \(t_i=t_j,\delta_i=1,\delta_j=0\). In the context of survival analysis, the risk scores of a comparable pair are said to be concordant with the survival outcomes if \(r_i>r_j\). The C-index is defined as the proportion of the concordant pairs among the comparable pairs. For comparable pair satisfying \(t_i<t_j,\delta_i=1\), we count 0.5 in the numerator of the concordance index for tied risk scores (\(r_i=r_j\)).

References

Asano, J., & Hirakawa, A. (2017). Assessing the prediction accuracy of a cure model for censored survival data with long-term survivors: Application to breast cancer data. Journal of Biopharmaceutical Statistics, 27(6), 918--932.

Harrell, F. E., Lee, K. L., & Mark, D. B. (1996). Multivariable prognostic models: Issues in developing models, evaluating assumptions and adequacy, and measuring and reducing errors. Statistics in medicine, 15(4), 361--387.

Examples

## See examples of function 'cox_cure'.