Auxiliary for Controlling iCoxph Fitting

Auxiliary function for iCoxph that enable users to specify the control parameters of the model estimation procedure. Internally, the arguments cm_gradtol, cm_stepmax, cm_steptol, and cm_max_iter are passed to function nlm as gradtol, stepmax, steptol, and iterlim, respectively.

iCoxph.control(tol_beta = 1e-6, tol_pi = 1e-8, cm_gradtol = 1e-6,
               cm_stepmax = 1e2, cm_steptol = 1e-6, cm_max_iter = 1e2,
               ecm_max_iter = 2e2, ...)

Arguments

tol_beta
A positive value specifying the tolerance that concludes the convergence of the covariate coefficient estimates. The tolerance is compared with the relative change between the estimates from two consecutive iterations that is measured by ratio of the L2-norm of their difference to the sum of their L2-norm. The default value is 1e-6.

tol_pi
A positive value specifying the tolerance that concludes the convergence of the probability estimates of uncertain records being true. The tolerance is compared with the relative change between the estimates from two consecutive iterations measured by ratio of L2-norm of their difference to the L2-norm of their sum. The default value is 1e-8.

cm_gradtol
A positive scalar giving the tolerance at which the scaled gradient is considered close enough to zero to terminate CM steps. The default value is 1e-6.

cm_stepmax
A positive scalar that gives the maximum allowable scaled step length in CM steps. The default value is 1e2.

cm_steptol
A positive scalar providing the minimum allowable relative step length in CM step. The default value is 1e-6.

cm_max_iter
A positive integer specifying the maximum number of iterations to be performed before each CM step is terminated. The default value is 1e2.

ecm_max_iter
A positive integer specifying the maximum number of iterations to be performed before the ECM algorithm is terminated. The default value is 2e2.

...
Other arguments for future usage. A warning will be thrown if any invalid argument is specified.

Value

A list of class intsurv-iCoxph.control containing all specified control parameters.

See also

Examples

## See examples of function 'iCoxph'.